+++ CHANGELOG.rst
@@ -195,6 +195,8 @@ Added
   again.
 * Add electric scenario parameters for eGon100RE
   `#699 <https://github.com/openego/eGon-data/issues/699>`_
+* Introduce Sanity checks for eGon2035 
+  `#628 <https://github.com/openego/eGon-data/issues/382>`_
 * Provide wrapper to catch DB unique violation
   `#514 <https://github.com/openego/eGon-data/issues/514>`_